improvement: apply suggestions from code review
