+++ crates/mun/test/main.mun
@@ -1,8 +1,3 @@
-// function to add two floats
-//fn add(a:int, b:int):int {
-//  a+b
-//}
-
 // function to subtract two floats
 fn subtract(a:float, b:float):float {
   a-b
