+++ crates/mun_codegen/src/code_gen/symbols.rs
@@ -1,10 +1,9 @@
+use super::abi_types::{gen_abi_types, AbiTypes};
 use crate::ir::dispatch_table::DispatchTable;
 use crate::ir::function;
 use crate::values::{BasicValue, GlobalValue};
 use crate::IrDatabase;
 use inkwell::attributes::Attribute;
-use inkwell::context::ContextRef;
-use inkwell::types::{ArrayType, IntType, StructType};
 use inkwell::values::{IntValue, PointerValue, UnnamedAddress};
 use inkwell::{
     module::{Linkage, Module},
@@ -50,6 +49,7 @@ pub fn type_info_query(_db: &impl IrDatabase, ty: Ty) -> TypeInfo {
     }
 }
 
+/// Construct an IR `MunTypeInfo` struct value for the specified `TypeInfo`
 fn type_info_ir(ty: &TypeInfo, module: &Module) -> StructValue {
     let context = module.get_context();
     let guid_values: [IntValue; 16] =
@@ -72,140 +72,42 @@ fn intern_string(module: &Module, str: &str) -> PointerValue {
     gen_global(module, &value, ".str").as_pointer_value()
 }
 
-struct AbiTypes {
-    pub guid_type: ArrayType,
-    pub privacy_type: IntType,
-    pub type_info_type: StructType,
-    pub function_signature_type: StructType,
-    pub function_info_type: StructType,
-    pub module_info_type: StructType,
-    pub dispatch_table_type: StructType,
-    pub assembly_info_type: StructType,
-}
-
-/// Returns an `AbiTypes` struct that contains references to all LLVM ABI types.
-fn gen_abi_types(context: ContextRef) -> AbiTypes {
-    let str_type = context.i8_type().ptr_type(AddressSpace::Const);
-
-    // Construct the `MunGuid` type
-    let guid_type = context.i8_type().array_type(16);
-
-    // Construct the `MunPrivacy` enum
-    let privacy_type = context.i8_type();
-
-    // Construct the `MunTypeInfo` struct
-    let type_info_type = context.opaque_struct_type("struct.MunTypeInfo");
-    type_info_type.set_body(
-        &[
-            guid_type.into(), // guid
-            str_type.into(),  // name
-        ],
-        false,
-    );
-
-    // Construct the `MunFunctionSignature` type
-    let function_signature_type = context.opaque_struct_type("struct.MunFunctionSignature");
-    function_signature_type.set_body(
-        &[
-            str_type.into(),                                     // name
-            type_info_type.ptr_type(AddressSpace::Const).into(), // arg_types
-            type_info_type.ptr_type(AddressSpace::Const).into(), // return_type
-            context.i16_type().into(),                           // num_arg_types
-            privacy_type.into(),                                 // privacy
-        ],
-        false,
-    );
-
-    // Construct the `MunFunctionInfo` struct
-    let function_info_type = context.opaque_struct_type("struct.MunFunctionInfo");
-    function_info_type.set_body(
-        &[
-            function_signature_type.into(), // signature
-            context
-                .void_type()
-                .fn_type(&[], false)
-                .ptr_type(AddressSpace::Const)
-                .into(), // fn_ptr
-        ],
-        false,
-    );
-
-    // Construct the `MunModuleInfo` struct
-    let module_info_type = context.opaque_struct_type("struct.MunModuleInfo");
-    module_info_type.set_body(
-        &[
-            str_type.into(),                                         // path
-            function_info_type.ptr_type(AddressSpace::Const).into(), // functions
-            context.i32_type().into(),                               // num_functions
-        ],
-        false,
-    );
-
-    // Construct the `MunDispatchTable` struct
-    let dispatch_table_type = context.opaque_struct_type("struct.MunDispatchTable");
-    dispatch_table_type.set_body(
-        &[
-            function_signature_type.ptr_type(AddressSpace::Const).into(), // signatures
-            context
-                .void_type()
-                .fn_type(&[], false)
-                .ptr_type(AddressSpace::Generic)
-                .ptr_type(AddressSpace::Const)
-                .into(), // fn_ptrs
-            context.i32_type().into(),                                    // num_entries
-        ],
-        false,
-    );
-
-    // Construct the `MunAssemblyInfo` struct
-    let assembly_info_type = context.opaque_struct_type("struct.MunAssemblyInfo");
-    assembly_info_type.set_body(
-        &[
-            module_info_type.into(),
-            dispatch_table_type.into(),
-            str_type.ptr_type(AddressSpace::Const).into(),
-            context.i32_type().into(),
-        ],
-        false,
-    );
-
-    AbiTypes {
-        guid_type,
-        privacy_type,
-        type_info_type,
-        function_signature_type,
-        function_info_type,
-        module_info_type,
-        dispatch_table_type,
-        assembly_info_type,
-    }
-}
-
+/// Construct a `MunFunctionSignature` struct for the specified HIR function.
 fn gen_signature_from_function<D: IrDatabase>(
     db: &D,
     module: &Module,
     types: &AbiTypes,
-    function: hir::Function,
+    function: &hir::Function,
 ) -> StructValue {
-    // Intern the name of the function
     let name_str = intern_string(&module, &function.name(db).to_string());
+    let ret_type_ir = gen_signature_return_type(db, module, types, function);
+    let params_type_ir = gen_signature_argument_types(db, module, types, function);
 
-    // Get the return value type
+    let body = function.body(db);
+    types.function_signature_type.const_named_struct(&[
+        name_str.into(),
+        params_type_ir.into(),
+        ret_type_ir.into(),
+        module
+            .get_context()
+            .i16_type()
+            .const_int(body.params().len() as u64, false)
+            .into(),
+        module.get_context().i8_type().const_int(0, false).into(),
+    ])
+}
+
+/// Given a function, construct a pointer to a `MunTypeInfo[]` global that represents the argument
+/// types of the function; or `null` if the function has no arguments.
+fn gen_signature_argument_types<D: IrDatabase>(
+    db: &D,
+    module: &Module,
+    types: &AbiTypes,
+    function: &hir::Function,
+) -> PointerValue {
     let body = function.body(db);
     let infer = function.infer(db);
-    let ret_type = infer[body.body_expr()].clone();
-    let ret_type_ir: PointerValue = if ret_type.is_empty() {
-        types
-            .type_info_type
-            .ptr_type(AddressSpace::Const)
-            .const_null()
-    } else {
-        let ret_type_const = type_info_ir(&db.type_info(ret_type), &module);
-        gen_global(module, &ret_type_const, "").as_pointer_value()
-    };
-
-    // Get the argument types
-    let params_type_ir: PointerValue = if body.params().is_empty() {
+    if body.params().is_empty() {
         types
             .type_info_type
             .ptr_type(AddressSpace::Const)
@@ -219,19 +121,29 @@ fn gen_signature_from_function<D: IrDatabase>(
                 .collect::<Vec<StructValue>>(),
         );
         gen_global(module, &params_type_array_ir, "").as_pointer_value()
-    };
+    }
+}
 
-    types.function_signature_type.const_named_struct(&[
-        name_str.into(),
-        params_type_ir.into(),
-        ret_type_ir.into(),
-        module
-            .get_context()
-            .i16_type()
-            .const_int(body.params().len() as u64, false)
-            .into(),
-        module.get_context().i8_type().const_int(0, false).into(),
-    ])
+/// Given a function, construct a pointer to a `MunTypeInfo` global that represents the return type
+/// of the function; or `null` if the return type is empty.
+fn gen_signature_return_type<D: IrDatabase>(
+    db: &D,
+    module: &Module,
+    types: &AbiTypes,
+    function: &hir::Function,
+) -> PointerValue {
+    let body = function.body(db);
+    let infer = function.infer(db);
+    let ret_type = infer[body.body_expr()].clone();
+    if ret_type.is_empty() {
+        types
+            .type_info_type
+            .ptr_type(AddressSpace::Const)
+            .const_null()
+    } else {
+        let ret_type_const = type_info_ir(&db.type_info(ret_type), &module);
+        gen_global(module, &ret_type_const, "").as_pointer_value()
+    }
 }
 
 /// Construct a global that holds a reference to all functions. e.g.:
@@ -251,7 +163,7 @@ fn gen_function_info_array<'a, D: IrDatabase>(
             value.set_linkage(Linkage::Private);
 
             // Generate the signature from the function
-            let signature = gen_signature_from_function(db, module, types, *f);
+            let signature = gen_signature_from_function(db, module, types, f);
 
             // Generate the function info value
             types.function_info_type.const_named_struct(&[
@@ -288,7 +200,7 @@ fn gen_dispatch_table<D: IrDatabase>(
     let signatures: Vec<StructValue> = dispatch_table
         .entries()
         .iter()
-        .map(|f| gen_signature_from_function(db, module, types, *f))
+        .map(|f| gen_signature_from_function(db, module, types, f))
         .collect();
 
     // Construct an IR array from the signatures
@@ -337,10 +249,6 @@ pub(super) fn gen_reflection_ir(
     dispatch_table: &DispatchTable,
     module: &Module,
 ) {
-    let context = module.get_context();
-    let str_type = context.i8_type().ptr_type(AddressSpace::Const);
-    let target = db.target();
-
     // Get all the types
     let abi_types = gen_abi_types(module.get_context());
 
@@ -360,6 +268,33 @@ pub(super) fn gen_reflection_ir(
     // Construct the dispatch table struct
     let dispatch_table = gen_dispatch_table(db, &abi_types, module, dispatch_table);
 
+    // Construct the actual `get_info` function
+    gen_get_info_fn(db, &module, &abi_types, module_info, dispatch_table);
+}
+
+/// Construct the actual `get_info` function.
+fn gen_get_info_fn(
+    db: &impl IrDatabase,
+    module: &&Module,
+    abi_types: &AbiTypes,
+    module_info: StructValue,
+    dispatch_table: StructValue,
+) {
+    let context = module.get_context();
+    let target = db.target();
+    let str_type = context.i8_type().ptr_type(AddressSpace::Const);
+
+    // Construct the return type of the `get_info` method. Depending on the C ABI this is either the
+    // `MunAssemblyInfo` struct or void. On windows the return argument is passed back to the caller
+    // through a pointer to the return type as the first argument. e.g.:
+    // On Windows:
+    // ```c
+    // void get_info(MunModuleInfo* result) {...}
+    // ```
+    // Whereas on other platforms the signature of the `get_info` function is:
+    // ```c
+    // MunModuleInfo get_info() { ... }
+    // ```
     let get_symbols_type = if target.options.is_like_windows {
         context.void_type().fn_type(
             &[abi_types
@@ -386,6 +321,8 @@ pub(super) fn gen_reflection_ir(
     let body_ir = db.context().append_basic_block(&get_symbols_fn, "body");
     builder.position_at_end(&body_ir);
 
+    // Get a pointer to the IR value that will hold the return value. Again this differs depending
+    // on the C ABI.
     let result_ptr = if target.options.is_like_windows {
         get_symbols_fn
             .get_nth_param(0)
@@ -402,7 +339,7 @@ pub(super) fn gen_reflection_ir(
     let num_dependencies_addr =
         unsafe { builder.build_struct_gep(result_ptr, 3, "num_dependencies") };
 
-    // Fill them with values
+    // Assign the struct values one by one.
     builder.build_store(symbols_addr, module_info);
     builder.build_store(dispatch_table_addr, dispatch_table);
     builder.build_store(
@@ -414,11 +351,13 @@ pub(super) fn gen_reflection_ir(
         context.i32_type().const_int(0 as u64, false),
     );
 
+    // Construct the return statement of the function.
     if target.options.is_like_windows {
         builder.build_return(None);
     } else {
         builder.build_return(Some(&builder.build_load(result_ptr, "")));
     }
 
+    // Run the function optimizer on the generate function
     function::create_pass_manager(&module, db.optimization_lvl()).run_on(&get_symbols_fn);
 }
