+++ crates/mun_codegen/src/lib.rs
@@ -1,11 +1,4 @@
 /// This library generates machine code from HIR using inkwell which is a safe wrapper around LLVM.
-
-#[macro_use]
-extern crate lazy_static;
-
-#[macro_use]
-extern crate failure;
-
 mod code_gen;
 mod db;
 mod ir;
